N:M relation

+++ django_relation/articles/views.py
@@ -3,10 +3,9 @@ from django.shortcuts import render, redirect, get_object_or_404
 # django가 주는 views에서 쓸 decorators http를 위한
 from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
-from IPython import embed
-from django.contrib import messages
 
 # Create your views here.
 def index(request):
@@ -32,15 +31,15 @@ def detail(request, article_pk):
 @login_required
 def create(request):
     if request.method == "POST":
-        article_form = ArticleForm(request.POST, request.FILES)
-        if article_form.is_valid():
-            article = article_form.save(commit=False)
+        form = ArticleForm(request.POST, request.FILES)
+        if form.is_valid():
+            article = form.save(commit=False)
             article.user = request.user
             article.save()
-            messages.success(request, '게시글 작성 성공')
+            messages.success(request, '게시글 작성 완료!!!!!')
             return redirect('articles:detail', article.pk)
-        else: 
-            messages.error(request, '너 잘못된 데이터를 넣었어!')
+        else:
+            messages.error(request, '너 잘못된 데이터를 넣었어!!!')
     else:
         form = ArticleForm()
     context = {
@@ -79,24 +78,56 @@ def delete(request, article_pk):
 
 @require_POST
 def comment_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.user = request.user
-        # comment.article_id = article.pk
-        comment.save()
-        return redirect('articles:detail', article.pk)
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+            return redirect('articles:detail', article.pk)
+        else:
+            context = {
+                'comment_form': comment_form,
+                'article': article
+            }
+            return render(request, 'articles/index.html', context)
     else:
-        context = {
-            'comment_form' : comment_form,
-            'article' : article
-        }
-    return redirect('articles:detail', context)
+        return redirect('accounts:login')
+    
 
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        if comment.user == request.user:
+            comment.delete()
+        return redirect('articles:detail', article_pk)
+    else:
+        return redirect('accounts:login')
+
+@login_required
+def like(request, article_pk):
+    # 특정 게시물에 대한 정보
+    article = get_object_or_404(Article, pk=article_pk)
+    # 좋아요를 누른 유저에 대한 정보
+    user = request.user
+    # 사용자가 게시글의 좋아요 목록에 있으면
+    if user in article.like_users.all():
+        article.like_users.remove(user)
+    else:
+        article.like_users.add(user)
+    return redirect('articles:index')
+
+@login_required
+def recommend(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    user = request.user
+    if user in article.recommend_users.all():
+        article.recommend_users.remove(user)
+    else:
+        article.recommend_users.add(user)
+    return redirect('articles:index')
+
+
